feat/RSS-ECOMM-3_18: addresses refactor
