+++ src/pages/Profile/components/ProfileAddressesForm.tsx
@@ -136,7 +136,7 @@ export function ProfileAddressesForm({
                     onSubmitRemove={handleRemoveAddress}
                     isValid={formState.isValid}
                     AddressComponent={
-                      // Transfet userAddressComponent to modal
+                      // Transfert userAddressComponent to modal
                       <UserAddressComponent
                         disabled={isLoading || isBusy}
                         isReadonly={isLoading || isBusy}
